+++ src/main/java/com/taskmanager/org/dto/TaskDTO.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 
 public class TaskDTO {
 
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Status status;
@@ -15,14 +15,14 @@ public class TaskDTO {
     private Category categoryId;
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
-    private Integer userId;
+    private Long userId;
 
     public TaskDTO() {}
 
-    public TaskDTO(Integer id, String title, String description, Status status,
+    public TaskDTO(Long id, String title, String description, Status status,
                    LocalDateTime dueDate, Category categoryId,
                    LocalDateTime createdAt, LocalDateTime updatedAt,
-                   Integer userId) {
+                   Long userId) {
         this.id = id;
         this.title = title;
         this.description = description;
@@ -35,8 +35,8 @@ public class TaskDTO {
     }
 
 
-    public Integer getId() { return id; }
-    public void setId(Integer id) { this.id = id; }
+    public Long getId() { return id; }
+    public void setId(Long id) { this.id = id; }
 
     public String getTitle() { return title; }
     public void setTitle(String title) { this.title = title; }
@@ -59,6 +59,6 @@ public class TaskDTO {
     public LocalDateTime getUpdatedAt() { return updatedAt; }
     public void setUpdatedAt(LocalDateTime updatedAt) { this.updatedAt = updatedAt; }
 
-    public Integer getUserId() { return userId; }
-    public void setUserId(Integer userId) { this.userId = userId; }
+    public Long getUserId() { return userId; }
+    public void setUserId(Long userId) { this.userId = userId; }
 }
