+++ src/main/java/com/taskmanager/org/model/Category.java
@@ -10,7 +10,7 @@ public class Category {
 
     @Id
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
-    private Integer id;
+    private Long id;
     @Column(nullable = false)
     private String name;
     @Column(nullable = false)
@@ -19,12 +19,12 @@ public class Category {
     public Category() {}
 
 
-    public Category(Integer id, String name, String color) {
+    public Category(Long id, String name, String color) {
         this.id = id;
         this.name = name;
         this.color = color;
     }
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
@@ -36,7 +36,7 @@ public class Category {
         return color;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
